# use a single map in the incoming streams map