Turns out that map access are not free if you do them in large numbers:

```
❯ benchstat old.txt new.txt                          
name                              old time/op  new time/op  delta
StreamsMap5000Stream0Closed-16    41.5ns ± 5%  39.8ns ± 2%   -4.11%  (p=0.000 n=15+15)
StreamsMap1000Stream0Closed-16    37.7ns ± 5%  33.8ns ± 2%  -10.45%  (p=0.000 n=13+15)
StreamsMap1000Stream10Closed-16   47.9ns ± 9%  35.1ns ± 7%  -26.88%  (p=0.000 n=16+16)
StreamsMap1000Stream600Closed-16  41.5ns ± 7%  32.9ns ±15%  -20.71%  (p=0.000 n=16+16)
StreamsMap1000Stream900Closed-16  37.3ns ± 6%  32.5ns ± 5%  -12.94%  (p=0.000 n=16+15)
StreamsMap100Stream0Closed-16     29.0ns ± 2%  27.3ns ± 8%   -5.97%  (p=0.000 n=15+16)
StreamsMap100Stream10Closed-16    43.8ns ± 9%  27.9ns ± 8%  -36.24%  (p=0.000 n=16+16)
StreamsMap100Stream60Closed-16    33.3ns ± 9%  28.2ns ± 8%  -15.18%  (p=0.000 n=16+16)
StreamsMap100Stream90Closed-16    27.9ns ± 2%  29.0ns ± 7%     ~     (p=0.090 n=15+16)
StreamsMap30Stream0Closed-16      28.4ns ± 2%  26.9ns ± 2%   -5.30%  (p=0.000 n=16+16)
StreamsMap30Stream5Closed-16      33.9ns ± 9%  29.5ns ±14%  -12.98%  (p=0.000 n=16+16)
```